requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Cors

pymysql
mysqlclient

requests
python-dotenv

gunicorn

sentence-transformers
chromadb
google-generativeai
gtts

ultralytics

Pillow
numpy
scikit-learn
opencv-python
torch